// types.ts
export interface Message {
  id: string;
  created_at: string;
  sender_id: string;
  receiver_id: string;
  content: string | null;
  image_url: string | null;
  audio_url: string | null;
  read: boolean;
}